+++ src/read_write/codec.rs
@@ -16,6 +16,11 @@ use crate::math;
 use cgmath::{BaseFloat, Vector3, Zero};
 use num::clamp;
 
+pub enum Encoding {
+    Plain,
+    AABBScaled(Vector3<f64>, f64, crate::octree::PositionEncoding),
+}
+
 pub fn fixpoint_encode<T>(value: f64, min: f64, edge_length: f64) -> T
 where
     T: num_traits::PrimInt + num_traits::Bounded + num_traits::NumCast,
